requirements.txt
Django==4.1.13
mysqlclient==2.*
requests==2.31.0
simplejson==3.*
mybatis_mapper2sql==0.1.9
django-auth-ldap==4.1.0
python-dateutil==2.8.1
pymysql==0.9.3
mysql-replication==0.22
django-redis==5.2.0
redis==5.0.3
gunicorn==20.0.4
pyecharts==1.9.1
aliyun-python-sdk-rds==2.1.1
supervisor==4.1.0
django-mirage-field==1.4.0
schema-sync==0.9.7
parsedatetime==2.4
sshtunnel==0.1.5
pycryptodome==3.20.0
pandas==1.5.*
clickhouse-driver==0.*
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-filter==21.1
drf-spectacular==0.22.0
pyotp==2.6.0
pillow==10.*
qrcode==7.3.1
django-environ==0.8.1
alibabacloud_dysmsapi20170525==2.0.9
aliyun-python-sdk-r-kvstore==2.20.8
aliyun-python-sdk-dds==3.7.14
tencentcloud-sdk-python-common==3.0.986
tencentcloud-sdk-python-cdb==3.0.986
tencentcloud-sdk-python-vpc==3.0.986
tencentcloud-sdk-python-redis==3.0.986
tencentcloud-sdk-python-mongodb==3.0.986
huaweicloudsdkcore==3.1.60
huaweicloudsdkdcs==3.1.60
huaweicloudsdkdds==3.1.60
huaweicloudsdkrds==3.1.60
huaweicloudsdkvpc==3.1.60
xlsxwriter==3.1.2
cryptography==42.*
environ==1.0